scripts/requirements-capture.txt
undetected-chromedriver>=3.5.0
selenium>=4.0.0
Pillow>=9.0.0
